requirements.txt
anthropic>=0.18.0
pillow>=10.0.0
requests>=2.31.0
replicate>=0.22.0
fal-client>=0.4.0
flask>=3.0.0
python-dotenv>=1.0.0
